+++ kb_luckyrich_backend/src/main/java/mul/cam/e/service/MyAssetService.java
@@ -87,7 +87,7 @@ public class MyAssetService {
         return totalValue;
     }
 
-    public int totalRealestate(int userId) {
+    public long totalRealestate(int userId) {
         return myAssetDao.totalRealestate(userId);
     }
